aws: Start making terraform changes for IPv6
If you run "create cluster", the VPC will have an IPv6 subnet assigned
by AWS (a /56).  Then, each subnet created in the VPC will get a slice
of this (a /64).

I can see that the VMs created all had IPv6 addresses associated with
them, at least from the AWS perspective.

The networking section of my current install-config looks like this:

  networking:
    clusterNetwork:
    - cidr: fd01::/48
      hostPrefix: 64
    machineCIDR: 10.0.0.0/16
    networkType: OVNKubernetes
    serviceNetwork:
    - fd02::/112

I kept machineCIDR as IPv4, since on AWS we can't get IPv6 only.  We
get IPv4 + optionally IPv6.  We use machineCIDR to specify the IPv4
address range we carve up in the VPC.  For IPv6, AWS gives us the /56
that we carve up, so we don't need a subnet specified in the config.

For clusterNetwork and serviceNetwork, both of these are used
exclusively inside the cluster, so I changed them to be specified as
private IPv6 subnets.